EdgeDetection: add a cast for `qreal`
FossilOrigin-Name: 3814c091edf0a0ab74685825c9b85ca7089e02d9

+++ src/EdgeDetection.cxx
@@ -160,7 +160,7 @@ int EdgeList::blurRadius() const
 
 void EdgeList::setContrastFactor(qreal factor)
 {
-    d->contrastFactor = qMax(0.0, factor);
+    d->contrastFactor = qMax(static_cast<qreal>(0), factor);
 }
 
 qreal EdgeList::contrastFactor() const
